requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2.2
gunicorn>=20.1.0
python-dotenv>=1.0.0
paypalrestsdk==1.13.1

django-cors-headers==4.0.0
